Do not choke when parsing "\c"
The mime types XML file contains "\chapter" as a way of identifying
LaTeX files.  But parsing this entry causes get_matches to fail with
an error in the eval about an incomplete string literal.

The issue is that \cX is interpreted as Ctrl-X, so in particular

    eval('"\c"')

Is interpreted as

    eval('"' + ctrl-")

Where `ctrl-"` is a single character.

This is a bit of a hacky work-around, but I think \c is the only case
like this.